refactor(flourishes): move flourish types into global osc-ui types files
